refactor: passCoreEvalProposal
