+++ packages/synthetic-chain/package.json
@@ -17,6 +17,7 @@
     "test:xs": "exit 0"
   },
   "dependencies": {
+    "@endo/zip": "^1.0.0",
     "better-sqlite3": "^9.2.2",
     "tsx": "^3.12.8",
     "typescript": "^5.3.3"
